# Speed up cleanMolStereo